% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_edge_density.R
\name{adjacency_edge_density}
\alias{adjacency_edge_density}
\title{Edge Density Threshold}
\usage{
adjacency_edge_density(corr, edge.density = 0.05)
}
\arguments{
\item{corr}{correlation matrix}

\item{edge.density}{threshold value}
}
\description{
Retrieves the adjacency matrix from the correlation matrix by keeping only
the higher absolute correlations up to the desired graph link density.
}
