% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_p_adjust.R
\name{adjacency_p_adjust}
\alias{adjacency_p_adjust}
\title{Get adjacency with the p-adjusted values from correlations}
\usage{
adjacency_p_adjust(
  x,
  method = "pearson",
  adjust = "bonferroni",
  alpha = 0.01,
  keep.psych = FALSE
)
}
\arguments{
\item{x}{(Required) matrix of data.}

\item{method}{(Optional) correlation type, possible choices are "pearson",
"spearman" or "kendall" (default spearman).}

\item{adjust}{(Optional) multiple test choosen, possible choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none".}

\item{alpha}{(Optional) level of confidence interval (default 0.01).}

\item{keep.psych}{(Optional) logical paramater which indicates if return the
results of corr.test (default FALSE).}
}
\description{
User wrapper of \code{\link{corr.test}} to get the
adjacency from correlation matrix. The function apply a
threshold on correlation matrix using the adjusted p-value with multiple test
corrections.
}
