% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacteria}
\alias{bacteria}
\title{Collection of metagenomics soil sample sequenced with shoutgun tecnique
(only bacteria).}
\format{
\subsection{\code{bacteria}}{

A matrix with 125 rows/sample and 1280 columns/species:
}
}
\usage{
bacteria
}
\description{
Collection of metagenomics soil sample sequenced with shoutgun tecnique
(only bacteria).
}
\keyword{datasets}
