% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_signed_louvain.R
\name{cluster_signed_louvain}
\alias{cluster_signed_louvain}
\title{Get Communities of a Signed Weighted Graph}
\usage{
cluster_signed_louvain(graph)
}
\arguments{
\item{graph}{weighted undirected network belong to \code{\link{igraph}} class.}
}
\description{
The algorithm tries to find dense subgraph, also called
communities in graphs via optimization of a signed definition of modularity
score.
}
