% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_communities_palette.R
\name{generate_communities_palette}
\alias{generate_communities_palette}
\title{Generate communities color palette}
\usage{
generate_communities_palette(N, seed = 123)
}
\arguments{
\item{N}{(Required) integer with the wanted number if distinct color.}

\item{seed}{random seed for reproducibility.}
}
\description{
Create a named character vector with the color palette
associated with the communities. The function is based on the
function \code{\link{distinctColorPalette}}.
}
\examples{
generate_communities_palette(8)

}
