% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palette_communities.R
\name{generate_palette_communities}
\alias{generate_palette_communities}
\title{Generate communities palette color}
\usage{
generate_palette_communities(N)
}
\arguments{
\item{N}{(Required) integer with the wanted number if distinct color.}
}
\description{
Create a named character vector with the color palette
associated with the communities. The function is based on the
function \code{\link{distinctColorPalette}}.
}
\examples{
palette_communities(8)

}
