% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_palette_rank.R
\name{generate_palette_rank}
\alias{generate_palette_rank}
\title{Generate rank palette color}
\usage{
generate_palette_rank(taxaID)
}
\arguments{
\item{taxaID}{(Required) character vector with the taxonomic IDs.
For a number of unique elements greater than 40, the colors may no longer be
clearly distinguishable.}
}
\description{
Create a named character vector with the color palette
associated with the taxonomy classification. The function is based on the
function \code{\link{distinctColorPalette}}.
}
\examples{
palette_rank(c("Proteobacteria","Firmicutes","Proteobacteria","Bacteroidetes","Firmicutes","Proteobacteria"))

}
