% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_palette.R
\name{generate_taxonomy_palette}
\alias{generate_taxonomy_palette}
\title{Generate taxonomy color palette}
\usage{
generate_taxonomy_palette(taxaID, seed = 123)
}
\arguments{
\item{taxaID}{(Required) character vector with the taxonomic IDs.
For a number of unique elements greater than 40, the colors may no longer be
clearly distinguishable.}

\item{seed}{random seed for reproducibility.}
}
\description{
Create a named character vector with the color palette
associated with the taxonomy classification. The function is based on the
\code{\link{distinctColorPalette}}.
}
\examples{
generate_taxonomy_palette(c("Proteobacteria","Firmicutes","Proteobacteria","Bacteroidetes","Firmicutes","Proteobacteria"))

}
