% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_signed.R
\name{layout_signed}
\alias{layout_signed}
\title{Signed-Weighted Graph Layout}
\usage{
layout_signed(graph, seed = 123)
}
\arguments{
\item{graph}{weighted undirected network belong to \code{\link{igraph}} class.}

\item{seed}{random seed for reproducibility.}
}
\description{
Signed version of fruchterman reingold layout. The function take
into account only the positive links.
}
