% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Samples metadata}
\format{
\subsection{\code{metadata}}{

A data.frame with the experimental variables of the samples.
The columns contain the geographical sites and the date.
}
}
\usage{
metadata
}
\description{
Samples metadata
}
\keyword{datasets}
