% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomic Classification}
\format{
\subsection{\code{taxonomy}}{

A character matrix with 1280 rows/taxa and 7 columns/taxonomic rank.
It contains the taxonomy classification of the bacteria.
}
}
\usage{
taxonomy
}
\description{
Taxonomic Classification
}
\keyword{datasets}
