% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triu2mat.R
\name{triu2mat}
\alias{triu2mat}
\title{Upper triangular values into square matrix}
\usage{
triu2mat(v, diag = 0)
}
\arguments{
\item{v}{(Required) upper triangular elements.}

\item{diag}{(Optional) diagonal elements of resulting matrix (default 0)}
}
\description{
reorganize a vector containing the values of an upper triangular
matrix into a symmetric square matrix.
}
