% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_signed_louvain.R
\name{cluster_signed_louvain}
\alias{cluster_signed_louvain}
\title{Get Communities of a Signed Weighted Graph}
\usage{
cluster_signed_louvain(graph, OS = "Linux")
}
\arguments{
\item{graph}{weighted undirected network belong to \code{\link{igraph}} class.}

\item{OS}{string with the operating system running. Possible choices are
"Linux","Windows","Mac" (default "Linux").}
}
\value{
\code{\link{communities}} igraph object able to manage to communities
graph info. The unique difference from igraph routine is the treatment with
the isolated nodes. In this case all isolated nodes are classified in the
community \code{'0'} and not as different communities of size one
}
\description{
The algorithm tries to find dense subgraph, also called
communities via optimization of a signed definition of modularity
score.
}
